# Faster+safer use of unsafe package + non-unsafe fallback implementations