Pardon the unhelpful PR title; wasn't sure what to call this...

Removed the `unsafeBytesToString` function, instead encapsulating `unsafe` functionality in a separate set of files (`fastbytes*`), thus preventing misuse by client code. Included are fast implementations of `BytesEqualStr`, which takes one string and one []byte argument, and `BytesParseFloat` and `BytesParseInt`, which each take a `[]byte` argument. None of these functions let unsafe strings escape, thus improving overall safety.

The main benefit of this refactoring is the ability to have non-unsafe fallback versions of these functions for platforms without the `unsafe` package (such as Google App Engine). Such implementations are included in `fastbytessafe.go`.

A somewhat faster implementation of BytesEqualStr is included than was previously used in the codebase. Sadly, string equality checks are a small portion of overall `Get*()` cost now, so there is little net performance impact. A significantly faster implementation of BytesParseInt is also included, but has little net impact for the same reason.

No noticeable net change in the benchmark, however the code is now cleaner, safer, and more portable.

Note that this patch exports `Bytes{EqualStr,ParseInt,ParseFloat}`, as they may be of use to client code. This is far safer than the old method of exporting `GetUnsafeString`, as these new functions are safe.

**Benchmark before change**:
```
BenchmarkJsonParserLarge-8 	  100000	     62860 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserMedium-8	  500000	     10639 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserSmall-8 	10000000	       985 ns/op	       0 B/op	       0 allocs/op
```

**Benchmark after change**:
```
BenchmarkJsonParserLarge-8 	  100000	     62428 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserMedium-8	 1000000	     10779 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserSmall-8 	10000000	       996 ns/op	       0 B/op	       0 allocs/op
```