@daboyuka 2016-03-27
Note: this PR is not ready yet; I just wanted you to be aware of what I was working on. The actual PR consists of only the last two commits (the rest are my earlier PR commits, since this PR is dependent on some of them).