@daboyuka 2016-03-27
Oh yeah, and the "Simpler (faster?) true/false/null literal checking" one I haven't made a separate PR for yet.