@daboyuka 2016-03-28
Note: there may be some issue merging this with PR #28 at this line:

`value := unsafeBytesToString(data[offset : endOffset+end])` (line 226 on this branch)

The equivalent line in PR #28 should be selected in the merge.