@daboyuka 2016-03-28
Looks like there's also a conflict on line 124; use this version of the line: `strbytes.Equal(keys[level-1], data[keyBegin:keyEnd]) {`