@daboyuka 2016-03-28
Note: this PR also removes `GetUnsafeString` because of the restructuring of unsafe functions. It can be added back if that's the final decision, though I would at least recommend it go in the new `strbytes` package with a non-unsafe version for platforms without the `unsafe` package, based on build flags.