@buger 2016-04-01
Yeah, change is big, did not had time to write comment till now. 

So, main idea here is to provide compatibility with environments that do not support `unsafe` package, right? And i like this idea. 
What i do not like is having it as a separate package, it make contributing way harder (you need to rewrite paths to package during merge every time, and this is kinda additional dependency, even internal), i prefer to include those packages directly to `jsonparser` namespace. 
Also "moreUnsafe" sounds dangerous :) Do you know real difference between `unsafe` and `moreUnsafe` versions? If you can give me link it will be perfect. 

Note that this PR will be merged manually, so probably you do not need to make more changes here, just answer questions above.

Also when you provide benchmarks, do not forget add `-test.benchmem` option to show memory allocation, some `unsafe` usage was made to remove allocations, its not only about CPU. 

And as always thank you so much for all your contributions!