@daboyuka 2016-04-01
Sure, these are good questions, I'm happy to try to answer. First, though, I have one request: would you mind to let me update my PR before you merge so that it won't require a manual merge? I know this code well, so it's easier and safer for me to do the update. Once you're satisfied with the plan, I can make the changes and notify you here.

Compatibility with other environments is one main point. The other is to encapsulate the `unsafe` functionality as much as possible to limit the chances for bad bugs.

Moving everything into the main package seems fine to me. The only reason I did it that way was to enforce encapsulation, but there are other ways to do that (including just not using the internal functions). I can move it back out.

`moreUnsafe` is based on this StackOverflow post: http://stackoverflow.com/a/31484416. I've included an explanation of how it in comparison to `unsafe` below.

Thanks for the info about `-test.benchmem`, I didn't know that one. I'll be sure to use it in the future.

`moreUnsafe` explanation:
As you know, both slices and strings are stored as a small headers that each contain a pointer to the actual data. The layout of a slice header is (currently) `(pointer) (len) (cap)`, whereas the layout of a string header is `(pointer) (len)` (both `(pointer)` and (`len`) are of the same type and alignment in both cases).

The `unsafe` method operates by constructing a new string header using the same pointer as a slice header; this is quite cheap, as it only requires copying two values (pointer and length) to a new stack-allocated header. This utilizes `reflect` types `SliceHeader` and `StringHeader` to be sure everything is copied right, and so it will work no matter how the pointer/len/cap fields are arranged within the headers.

The `moreUnsafe` method operates on the assumption that (pointer) and (len) in both kinds of header 1) have the same types, 2) have the same alignments, and 3) appear in the same order (all of which are true since the beginning of Go until now, I think). Given these facts, `moreUnsafe` simply reinterprets the slice's header as a string header in-place: the (pointer) and (len) fields are where they are supposed to be, and the (cap) field is just ignored. This approach is slightly faster because it requires no operations of any kind (it's just a pointer reinterpretation). For short strings, this can be noticeably faster than `unsafe` (see the benchmark I added in that package).