@buger 2016-04-01
Can you imagine situation when moreUnsafe version breaks? For example length of string that contains unicode characters can differ from length of same []byte, can you make some unicode tests?

I'm ok if you will make this changes, but not that currently there is conflicts with current master, so maybe make sense to create new PR, or solve them first, here. Pls move all the functions to `jsonparser` namespace, and ensure that your changes does not add additional memory allocations. 

Thanks!