@daboyuka 2016-04-01
The length of a string that contains Unicode _cannot_ differ from the length of the same []byte. For example: http://play.golang.org/p/5lgSA2ko8-. The only way `moreUnsafe` would break is if Go changed the slice/string header structs in C, which hasn't happened yet, and could only happen in a new version of Go. Despite being called `moreUnsafe`, I think it's safe enough :).

I'll work on the merge soon, maybe later today (Friday). Thanks!