@buger 2016-04-01
I see, thanks! I thought len(string) will print actual length for runes... You have to use `utf8.RuneCountInString` to do it.

http://play.golang.org/p/ckqYkyJFCQ