@daboyuka 2016-04-01
For what it's worth, I've been paying attention to that distinction as I read the code, and so far I haven't seen any UTF-related bugs. That proves nothing, though :).