@daboyuka 2016-04-06
I reworked the branch to address your suggestions. You may find the updated benchmarks in the original post.

* Updated to the latest version of buger/jsonparser/master (no more merge conflicts)
* Removed the `strbytes` package and moved the files into the main `jsonparser` package
* Simplified the source code a good bit
* Developed a custom implementation of ParseInt that operates directly on a []byte and is significantly faster than the `strconv` even when ignoring string conversion (it gains speedup by hard-coding base 10 and 64-bit integer return).