@daboyuka 2016-04-06
Ok, my changes in this PR are finally finished. Let me know if there are any further issues.