@daboyuka 2016-04-07
Fixed some merge conflicts with respect to latest master, so this will merge again.