@daboyuka 2016-04-11
Note: if you merge #40 first, a block of code in `searchKeys` will conflict; here's what the conflict fix should look like:

```
if data[i] == ':' {
	key := data[keyBegin:keyEnd]

	// for unescape: if there are no escape sequences, this is cheap; if there are, it is a
	// bit more expensive, but causes no allocations unless len(key) > unescapeStackBufSize
	if keyUnesc, err := Unescape(key, stackbuf[:]); err != nil {
		return -1
	} else {
		if keyLevel == level-1 && // If key nesting level match current object nested level
			BytesEqualStr(&keyUnesc, keys[level-1]) {
			keyLevel++
			// If we found all keys in path
			if keyLevel == lk {
				return i + 1
			}
		}
	}
} else {
	i--
}
```