@buger 2016-04-13
Merged with some refactoring, thanks!